[Refactor] 병합 후 오류나는 부분 정리
