+++ src/main/java/project/moseup/controller/TeamPageController.java
@@ -6,16 +6,18 @@ import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
 import project.moseup.domain.TeamAskBoard;
+import project.moseup.dto.TeamAskForm;
 import project.moseup.service.TeamAskBoardService;
 import project.moseup.service.TeamService;
 
 @Controller
 @RequiredArgsConstructor
+@RequestMapping("/teams/")
 public class TeamPageController {
 
 	private final TeamService teamService;
