+++ src/main/resources/templates/admin/memberList.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html xmlns:th="http://www.thymeleaf.org">
 
-<head th:replace="fragments/adminCommon :: head('회원리스트')">
+<head th:replace="layouts/fragments/adminCommon :: head('회원리스트')">
     <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -16,7 +16,7 @@
 
 <body>
     <header class="navbar navbar-dark sticky-top bg-dark flex-md-nowrap p-0 shadow"
-        th:replace="fragments/adminCommon :: header">
+        th:replace="layouts/fragments/adminCommon :: header">
         <!-- <a class="navbar-brand col-md-3 col-lg-2 me-0 px-3" href="#">Moseup</a>
         <button class="navbar-toggler position-absolute d-md-none collapsed" type="button" data-bs-toggle="collapse"
             data-bs-target="#sidebarMenu" aria-controls="sidebarMenu" aria-expanded="false"
@@ -31,7 +31,7 @@
         </div> -->
     </header>
 
-    <div class="container-fluid" th:replace="fragments/adminCommon :: sidebar('memberList')">
+    <div class="container-fluid" th:replace="layouts/fragments/adminCommon :: sidebar('memberList')">
         <!-- <div class="row">
             <nav id="sidebarMenu" class="col-md-3 col-lg-2 d-md-block bg-light sidebar collapse">
                 <div class="position-sticky pt-3">
@@ -111,7 +111,7 @@
             </nav>
         </div> -->
     </div>                                              
-    <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4" th:replace="fragments/adminCommon :: title">
+    <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4" th:replace="layouts/fragments/adminCommon :: title">
         <!-- <div
             class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
             <h1 class="h2">제목 작성</h1>
