--- src/main/resources/templates/layout/layout.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lagn="ko" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-
-<!-- header -->
-<th:block th:replace="fragments/header :: header"></th:block>
-
-<!-- layout CSS -->
-<link href="/css/layout.css" rel="stylesheet" type="text/css" />
-
-<body>
-	<!-- navbar -->
-	<th:block th:replace="fragments/navbar :: navbar"></th:block>
-	<!-- 본문 -->
-	<th:block layout:fragment="content"></th:block>
-	<!-- floatingActionButton -->
-	<th:block th:replace="fragments/floatingActionButton :: floatingActionButton">
-</body>
-<!-- footer -->
-<th:block th:replace="fragments/footer :: footer">
-</html>
