+++ src/main/resources/templates/layouts/fragments/adminCommon.html
@@ -41,7 +41,7 @@
                     <ul class="nav flex-column">
                         <li class="nav-item">
                             <a class="nav-link" th:classappend="${sidebar} == 'dashBoard'? 'active'" href="#"
-                                th:href="@{/admin}">
+                                th:href="@{/templates/admin}">
                                 <span class="sr-only" aria-current="page" th:if="${sidebar} == 'dashBoard'"></span>
                                 대시보드
                             </a>
