+++ src/main/resources/templates/layouts/fragments/floatingActionButton.html
