+++ src/main/resources/templates/layouts/fragments/footer.html
@@ -1,21 +1,21 @@
 <html lagn="ko" xmlns:th="http://www.thymeleaf.org">
 <div th:fragment="footer">
 	<div class="footer-container footer-width">
-		<footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
-			<p class="col-md-4 mb-0 text-muted footer-left">&copy; 2022 Company, Inc</p>
+	  <footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
+	    <p class="col-md-4 mb-0 text-muted footer-left">&copy; 2022 Company, Inc</p>
 
-			<a href="/static" class="col-md-4 d-flex align-items-center justify-content-center mb-3 mb-md-0 me-md-auto link-dark text-decoration-none">
-				<img class="bi me-2" width="45" height="35" alt="moseup" src="/images/logo.png">
-			</a>
+	    <a href="/static" class="col-md-4 d-flex align-items-center justify-content-center mb-3 mb-md-0 me-md-auto link-dark text-decoration-none">
+	      <img class="bi me-2" width="45" height="35" alt="moseup" src="/images/logo.png">
+	    </a>
 
-			<ul class="nav col-md-4 justify-content-end footer-right">
-				<li class="nav-item"><a href="#" class="nav-link px-2 text-muted">Home</a></li>
-				<li class="nav-item"><a href="#" class="nav-link px-2 text-muted">Features</a></li>
-				<li class="nav-item"><a href="#" class="nav-link px-2 text-muted">Pricing</a></li>
-				<li class="nav-item"><a href="#" class="nav-link px-2 text-muted">FAQs</a></li>
-				<li class="nav-item"><a href="#" class="nav-link px-2 text-muted">About</a></li>
-			</ul>
-		</footer>
+	    <ul class="nav col-md-4 justify-content-end footer-right">
+	      <li class="nav-item"><a href="#" class="nav-link px-2 text-muted">Home</a></li>
+	      <li class="nav-item"><a href="#" class="nav-link px-2 text-muted">Features</a></li>
+	      <li class="nav-item"><a href="#" class="nav-link px-2 text-muted">Pricing</a></li>
+	      <li class="nav-item"><a href="#" class="nav-link px-2 text-muted">FAQs</a></li>
+	      <li class="nav-item"><a href="#" class="nav-link px-2 text-muted">About</a></li>
+	    </ul>
+	  </footer>
 	</div>
 </div>
 </html>
