+++ src/main/resources/templates/teams/askBoardWriteForm.html
@@ -1,40 +1,40 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns="http://www.w3.org/1999/xhtml"  xmlns:th="http://www.thymeleaf.org"
+	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layouts/layout}">
 <head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
-<link rel="stylesheet" th:href="@{/css/bootstrap.min.css}">
-<link rel="stylesheet" type="text/css" href="/css/team.css" />
-<script type="text/javascript" th:src="@{/js/jquery.js}"></script>
-<script type="text/javascript" th:src="@{/js/bootstrap.min.js}"></script>
+	<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+	<title>Insert title here</title>
+	<link rel="stylesheet" th:href="@{/css/bootstrap.min.css}">
+	<link rel="stylesheet" type="text/css" href="/css/team.css" />
+	<script type="text/javascript" th:src="@{/js/jquery.js}"></script>
+	<script type="text/javascript" th:src="@{/js/bootstrap.min.js}"></script>
 </head>
 <body>
-	<header th:replace="fragments/header :: header"></header>
-	<!-- 네비바 -->
-	<nav th:insert="fragments/navbar :: nav"></nav>
+<!-- 본문 코드 시작 -->
+<th:block layout:fragment="content">
 	<section id="feature">
 		<div class="container">
-		<h2 class="text-primary" align="center">문의 하기</h2>
-		<form th:action="@{teamAskBoardWriteForm/createTeamAsk}" th:object="${teamAsk}" method="post">
-			<div class="form-group">
-				<label th:for="nickname">작성자</label> 
-				<input type="text" th:field="*{nickname}" class="form-control" placeholder="별명을 입력하세요">
+			<h2 class="text-primary" align="center">문의 하기</h2>
+			<form th:action="@{teamAskBoardWriteForm/createTeamAsk}" th:object="${teamAsk}" method="post">
+				<div class="form-group">
+					<label th:for="nickname">작성자</label>
+					<input type="text" th:field="*{nickname}" class="form-control" placeholder="별명을 입력하세요">
+				</div>
+				<div class="form-group">
+					<label th:for="teamAskSubject">제목</label>
+					<input type="text" th:field="*{teamAskSubject}" class="form-control" placeholder="제목을 입력하세요">
+				</div>
+				<div class="form-group">
+					<label th:for="teamAskContent">내용</label>
+					<textarea rows="5" cols="40" th:field="*{teamAskContent}" class="form-control" placeholder="내용을 입력하세요"></textarea>
+				</div>
+				<button type="submit" class="btn btn-primary">등록</button>
+			</form>
+			<div>
+				<a href="/teamAskBoard" class="btn">목록</a>
 			</div>
-			<div class="form-group">
-				<label th:for="teamAskSubject">제목</label> 
-				<input type="text" th:field="*{teamAskSubject}" class="form-control" placeholder="제목을 입력하세요">
-			</div>
-			<div class="form-group">
-				<label th:for="teamAskContent">내용</label> 
-				<textarea rows="5" cols="40" th:field="*{teamAskContent}" class="form-control" placeholder="내용을 입력하세요"></textarea>
-			</div>
-			<button type="submit" class="btn btn-primary">등록</button>
-		</form>
-		<div>
-			<a href="/teamAskBoard" class="btn">목록</a>
-		</div>
 		</div>
 	</section>
+</th:block>
 </body>
 </html>
