+++ src/main/resources/templates/teams/teamAskBoardDetail.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns="http://www.w3.org/1999/xhtml"  xmlns:th="http://www.thymeleaf.org"
+	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layouts/layout}">
 <head>
-<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
-<link rel="stylesheet" type="text/css" href="/css/team.css" />
-<title>Insert title here</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+	<link rel="stylesheet" type="text/css" href="/css/team.css" />
+	<title>Insert title here</title>
 </head>
 <body>
-<header th:replace="fragments/header :: header"></header>
-	<!-- 네비바 -->
-	<nav th:insert="fragments/navbar :: nav"></nav>
+<!-- 본문 코드 시작 -->
+<th:block layout:fragment="content">
 	<div class="container" align="center">
 		<h2 class="text-primary">문의글 상세조회</h2>
 		<table class="table table-striped">
@@ -35,5 +35,6 @@
 		</table>
 		<a href="/teamAskBoard" class="btn btn-success">게시글 목록</a>
 	</div>
+</th:block>
 </body>
 </html>
